OracleBot-Pro/pyproject.toml
[build-system]
requires = ["setuptools>=67.8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oraclebot-pro"
version = "1.0.0"
description = "Enterprise-Grade Algorithmic Trading System with MT5 & OpenClaw Integration"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Proprietary"}

# Production dependencies
dependencies = [
    "MetaTrader5==5.0.5572",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "python-telegram-bot>=20.7",
    "aiohttp>=3.9.0",
    "pydantic>=2.5.0",
    "redis>=5.0.0",
    "tenacity>=8.2.0",
    "structlog>=23.0.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.6.0",
    "ruff>=0.1.0",
    "black>=23.11.0",
    "isort>=5.12.0"
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "B", "C4", "ERA", "RUF"]
ignore = ["E501", "ERA001", "I001", "W291", "W292", "W293"]

[tool.black]
line-length = 100

[tool.mypy]
python_version = "3.12"
strict = true
files = ["core"]
ignore_missing_imports = true
disable_error_code = ["import-untyped"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
